Fixed MacOSXS/OpenStep/NextStep compiler warnings.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@966 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/impexp/impexp.cpp
@@ -1905,9 +1905,9 @@ void converter::print_sizes ( ) {
   fprintf ( logfile,  "  unsigned int        %d       0       %u\n", 
     (int)sizeof ( unsigned int ), UINT_MAX );
   fprintf ( logfile,  "  long int            %d       %ld      %ld\n", 
-    (int)sizeof ( long int ), LONG_MIN, LONG_MAX );
+    (int)sizeof ( long int ), (long int)LONG_MIN, (long int)LONG_MAX );
   fprintf ( logfile,  "  unsigned long int   %d       0       %lu\n", 
-    (int)sizeof ( unsigned long int ), ULONG_MAX );
+    (int)sizeof ( unsigned long int ), (unsigned long int)ULONG_MAX );
 /*
   FLT_MIN, FLT_MAX, DBL_MIN, DBL_MAX not defined on Microsoft C.
 */
